Refactor {de,}serialization to be more lenient (#6)
Rather than using an enum of all possible types, just use a raw JSON
value to allow users more flexibility with implementing their own
request/response types. E.g., when a new field is added, or a new
request type is added, users can handle them in the application code
without needing to wait for a new version of this library to be
released.