Refactor request/response type enums (#20)
The previous approach in #19 had an issue where users could potentially
pattern match the `Other` variant and bypass the `PartialEq` checks for
ensuring the types were the same.

This PR removes the fallback variant of the request/response type enums
and instead adds a layer of indirection to disallow the user from
directly pattern matching the variant. It also makes it less verbose
to add support for similar enums (e.g., the hotkey types) in later PRs.