Use tower's `BoxCloneService`. Remove re-exports.
This service now exists in a released version of `tower` (previously we
were using a vendored version of it in-crate), so we can use it
directly. This is technically a breaking change, even though likely no
one was depending on the service directly, so bumping to 0.3.

Also removes some re-exports that were kept around for backwards
compatibility.